refactor: Controller, View 생성 및 역할 할당
