+++ src/main/java/com/mission2/racingcar/domain/CarRacingService.java
@@ -10,18 +10,6 @@ public class CarRacingService {
     public static final int INIT_SCORE = 1;
     public static final String CAR_NAME_DELIMITER = ",";
 
-    public void game(List<String> carNames, int gameCount) {
-        System.out.println("실행결과");
-
-        carNames.stream()
-                .map(name -> new Car(name, INIT_SCORE, new RandomMoveStrategy()))
-                .forEach(System.out::println);
-
-        Race race = initRace(gameCount, carNames);
-        proceedGame(race);
-        printWinner(race.getWinners());
-    }
-
     public boolean checkCarNamesNotOverMaxCarCount(List<String> carNames) {
         long count = carNames.stream().filter(car -> car.length() > MAX_CAR_NAME_COUNT).count();
 
@@ -47,7 +35,7 @@ public class CarRacingService {
 
     public Race initRace(int gameCount, List<String> carNames) {
         List<Car> carList = carNames.stream()
-                .map(carName -> new Car(carName, INIT_SCORE, new RandomMoveStrategy()))
+                .map(carName -> new Car(carName, INIT_SCORE, new ThresholdFourStrategy()))
                 .collect(Collectors.toList());
 
         return new Race
@@ -57,23 +45,4 @@ public class CarRacingService {
                 .build();
     }
 
-    public void proceedGame(Race race) {
-        for (int i = 0; i < race.getGameCount(); i++) {
-            System.out.println();
-            racing(race);
-        }
-    }
-
-    private void racing(Race race) {
-        for (Car car : race.getCars()) {
-            Random random = new Random();
-            car.raceByRandomNumber(random.nextInt(10));
-            System.out.println(car);
-        }
-    }
-
-    public void printWinner(String[] winners) {
-        String winner = String.join(",", winners);
-        System.out.println("\n" + winner + "가 최종 우승했습니다.");
-    }
 }
