+++ src/main/java/com/mission2/racingcar/domain/MoveStrategy.java
@@ -1,5 +1,5 @@
 package com.mission2.racingcar.domain;
 
 public interface MoveStrategy {
-    public abstract boolean forward(int input);
+    boolean forward(int input);
 }
