+++ src/main/java/com/mission2/racingcar/domain/ThresholdFourStrategy.java
@@ -1,6 +1,6 @@
 package com.mission2.racingcar.domain;
 
-public class RandomMoveStrategy implements MoveStrategy {
+public class ThresholdFourStrategy implements MoveStrategy {
     public static final int FORWARD_THRESHOLD = 4;
 
     @Override
