+++ src/main/java/com/mission2/racingcar/view/View.java
@@ -1,5 +1,7 @@
 package com.mission2.racingcar.view;
 
+import java.util.List;
+
 public class View {
 
     InputView inputView = new InputView();
@@ -9,5 +11,8 @@ public class View {
         return inputView.getUserInput(message);
     }
 
+    public void printOutput(List<String> carNames, int gameCount) {
+        outputView.printOutput(carNames, gameCount);
+    }
 
 }
