+++ src/test/java/com/mission2/racingcar/domain/CarTests.java
@@ -11,7 +11,7 @@ class CarTests {
     @DisplayName("Random 숫자의 값이 0~3 이면 전진하지 않는다")
     void Random_값_정지() {
         //given
-        Car car = new Car("TestCar", 0, new RandomMoveStrategy());
+        Car car = new Car("TestCar", 0, new ThresholdFourStrategy());
         int stopValues = 3;
 
         //when
@@ -25,7 +25,7 @@ class CarTests {
     @DisplayName("Random 숫자의 값이 4~9 이면 전진한다")
     void Random_값_전진() {
         //given
-        Car car = new Car("TestCar", 0, new RandomMoveStrategy());
+        Car car = new Car("TestCar", 0, new ThresholdFourStrategy());
         int forwardValues = 4;
 
         //when
