+++ src/test/java/com/mission2/racingcar/domain/RaceTests.java
@@ -13,8 +13,8 @@ class RaceTests {
     @DisplayName("종료시 우승자를 출력한다")
     void 종료시_우승자_출력() {
         //given
-        Car[] carList = {new Car("AAA", 3, new RandomMoveStrategy()), new Car("BBB", 3, new RandomMoveStrategy()),
-                new Car("CCC", 4, new RandomMoveStrategy()), new Car("DDD", 5, new RandomMoveStrategy())};
+        Car[] carList = {new Car("AAA", 3, new ThresholdFourStrategy()), new Car("BBB", 3, new ThresholdFourStrategy()),
+                new Car("CCC", 4, new ThresholdFourStrategy()), new Car("DDD", 5, new ThresholdFourStrategy())};
         Race race = new Race(5, Arrays.asList(carList));
         String[] winners = race.getWinners();
